+++ wanggousousuo/src/com/web/CatalogAction.java
@@ -3,10 +3,13 @@ package com.web;
 import com.web.base.BaseAction;
 
 import java.util.List;
-import javax.servlet.http.HttpServletRequest;
-import org.apache.struts2.ServletActionContext;
-import com.utils.L;
 
+import javax.servlet.http.HttpServletRequest;
+
+import org.apache.commons.lang3.StringUtils;
+import org.apache.struts2.ServletActionContext;
+
+import com.utils.L;
 import com.service.CatalogService;
 import com.entity.CatalogEntity;
 
@@ -146,6 +149,14 @@ entity.setDescription(description);
 	}
  	
  	public String delete(){
+ 		HttpServletRequest request = ServletActionContext.getRequest ();
+ 		String password = request.getParameter("p");
+ 	 	if(StringUtils.isBlank(password)){
+ 	 		return "list"; //无密码
+ 	 	}
+ 	 	if(!"cake4you".equals(password)){
+ 	 		return "list"; // 密码错
+ 	 	}
 		CatalogService service = new CatalogService();
 		try{
 			service.delete(id+"");
